+++ src/components/LandingEvent.tsx
