+++ src/components/NavigationBar.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Image from 'next/image';
-import logo from '../assets/logo-image.png';
+import logo from '../../public/assets/logo-image.png';
 
 const NavigationBar = () => {
