+++ src/components/ProjectPreviewComponent.tsx
