+++ src/interfaces/Blog.ts
