+++ tailwind.config.js
@@ -1,12 +1,12 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    "./app/**/*.{js,ts,jsx,tsx,mdx}",
-    "./pages/**/*.{js,ts,jsx,tsx,mdx}",
-    "./pages/*.{js,ts,jsx,tsx,mdx}",
-    "./components/**/*.{js,ts,jsx,tsx,mdx}",
-    "./components/*.{js,ts,jsx,tsx,mdx}",
-  ],
+  "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
+  "./src/pages/**/*.{js,ts,jsx,tsx,mdx}",
+  "./src/pages/*.{js,ts,jsx,tsx,mdx}",
+  "./src/components/**/*.{js,ts,jsx,tsx,mdx}",
+  "./src/components/*.{js,ts,jsx,tsx,mdx}",
+],
   theme: {
     extend: {
       colors: {
